code cleanup
git-svn-id: svn+ssh://svn.cern.ch/reps/atlasoff/PhysicsAnalysis/SUSYPhys/Factory/ZeroLeptonRun2/trunk@685140 4525493e-7705-40b1-a816-d608a930855b

+++ Root/FilterUpdateMerge.cxx
@@ -74,7 +74,7 @@ void  FilterUpdateMerge::process(TTree* outTree, const std::string& inTreeName,
     if ( doCRYVars) inCRYVars.setAddresses(tree);
 
     // loop over entries
-    for ( size_t j = 0; j < tree->GetEntries(); ++j ) {
+    for ( size_t j = 0; j < static_cast<size_t>(tree->GetEntries()); ++j ) {
       tree->GetEntry(j);
       outVars = inVars.ntv;
       outRTVars = inRTVars.RTntv;
